Move mongo into database package and make private only expose default method from database package to return default Repository
